code/requirements.txt
django-dotenv==1.4.2
gunicorn==20.1.0
psycopg2-binary==2.9.3

#Project Specific
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.24.0
beautifulsoup4==4.4.1
billiard==3.5.0.2
boto==2.49.0
boto3==1.7.72
botocore==1.10.72
celery==4.0.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coverage==3.7.1
cryptography==2.7
Cython==0.23.5
decorator==4.0.9
diff-match-patch==20110725.1
dj-pagination==2.4.0
Django==2.1.7
django-admin-tools==0.6.0
django-aggregate-if==0.5
django-appconf==1.0.1
django-autoslug==1.9.4
django-bootstrap3-datetimepicker==2.2.3
django-compat==1.0.15
django-coverage==1.2.4
django-debug-panel==0.8.2
django-debug-toolbar==1.4
django-discover-runner==1.0
django-extensions==2.1.6
django-fabric==2.1.0
django-filter==2.1.0
django-hijack==2.1.10
django-ipware==1.1.6
django-kombu==0.9.4
django-model-utils==2.3.1
django-nose==1.4
django-notifications-hq==0.7.0
django-pytest==0.2.0
django-redis==4.10.0
django-role-permissions==2.0.1
django-ses==2.0.0
django-shortuuidfield==0.1.3
django-sitetree==1.12.0
django-smuggler==0.9.1
django-storages==1.5.2
django-webpack-loader==0.4.1
django-webtest==1.7.8
djangorestframework==3.9.1
docutils==0.14
dreampie==1.2.1
ecdsa==0.13
et-xmlfile==1.0.1
Fabric==1.10.2
funcsigs==0.4
future==0.17.1
fuzzywuzzy==0.17.0
gcm==0.3
idna==2.8
invoke==0.10.1
ipython==4.1.2
ipython-genutils==0.1.0
jdcal==1.3
jmespath==0.9.3
jsonfield==1.0.3
kombu==4.0.2
kombu-sqlalchemy==1.1.0
mock==1.3.0
model-mommy==1.6.0
mysqlclient==1.3.10
nose==1.3.6
ntlm-auth==1.5.0
numpy==1.14.0
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.22.0
paramiko==1.16.0
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
phonetics==1.0.5
pickleshare==0.6
Pillow==4.1.1
ptyprocess==0.5.1
py==1.4.34
pycparser==2.19
pycrypto==2.6.1
pyOpenSSL==19.0.0
pyparsing==2.2.0
pytest==3.1.2
pytest-django==3.1.2
python-dateutil==2.6.1
python-http-client==3.0.0
python-Levenshtein==0.12.0
python-magic==0.4.13
pytz==2018.3
redis==3.2.0
requests==2.22.0
requests-ntlm2==6.2.5
s3transfer==0.1.13
selenium==3.141.0
sendgrid==5.3.0
sentry-sdk==1.5.11
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.11
sqlparse==0.1.19
suds-jurko==0.6
traitlets==4.1.0
urllib3==1.23
uWSGI==2.0.17.1
vine==1.1.3
vobject==0.9.5
waitress==0.8.10
WebOb==1.5.1
WebTest==2.0.20
xlrd==1.0.0
XlsxWriter==1.0.0
xlwt==1.2.0
